split EventsDash into specialized files
